getting_to_the_grit

+++ libft/src/libc/ft_dtoa.c
@@ -6,7 +6,7 @@
 /*   By: mmosk <mmosk@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2024/01/15 12:27:06 by mmosk         #+#    #+#                 */
-/*   Updated: 2024/01/15 21:44:56 by mmosk         ########   odam.nl         */
+/*   Updated: 2024/01/16 22:11:44 by mmosk         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ char	*ft_dtoa(double in, char *out, int length)
 			past_point = 0;
 		}
 		out[length + past_point] =
-			'0' + (size_t)(in * pow(10, length - digits)) % 10;
+			'0' + (unsigned int)(in * pow(10, length - digits)) % 10;
 	}
 	return (out);
 }
